example/requirements.txt
django-nose==0.1.3
